GlusterFS: Remove extra options
The APB is an opinionated deployment and assumes things about the
underlying OCP environment.  As such, the default values for gluster
namespace, heketi route and secret, and whether to use an external
provisioner do not need to be changed.  Remove these from the apb to
provide a clean install experience.

Signed-off-by: Adam Litke <alitke@redhat.com>